feat: Replace getValue check with getVariable for variable fields.
Getting a variable using getValue applies only to FieldVariable. There
was a note to indicate this; this replaces it with the relevant check.

Also make getVariable available on fields, matching the existing referencesVariables/refreshVariableName prototypes.

This ensures that variables can be used with fields other than
dropdowns, as seems to be the intention.